Move userprofilepicturelist to a better spot
