+++ packages/web/src/components/notification/Notification/CommentReactionNotification.tsx
@@ -10,6 +10,7 @@ import { CommentReactionNotification as CommentReactionNotificationType } from '
 import { IconMessage } from '@audius/harmony'
 import { useDispatch } from 'react-redux'
 
+import { UserProfilePictureList } from 'components/user-profile-picture-list'
 import { useIsMobile } from 'hooks/useIsMobile'
 import { make, track } from 'services/analytics'
 import {
@@ -26,7 +27,6 @@ import { NotificationHeader } from './components/NotificationHeader'
 import { NotificationTile } from './components/NotificationTile'
 import { OthersLink } from './components/OthersLink'
 import { UserNameLink } from './components/UserNameLink'
-import { UserProfilePictureList } from './components/UserProfilePictureList'
 import { entityToUserListEntity, USER_LENGTH_LIMIT } from './utils'
 
 const messages = {
