+++ packages/web/src/components/notification/Notification/RepostNotification.tsx
@@ -7,6 +7,7 @@ import {
 } from '@audius/common/store'
 import { useDispatch } from 'react-redux'
 
+import { UserProfilePictureList } from 'components/user-profile-picture-list'
 import { useIsMobile } from 'hooks/useIsMobile'
 import {
   setUsers as setUserListUsers,
@@ -22,7 +23,6 @@ import { NotificationHeader } from './components/NotificationHeader'
 import { NotificationTile } from './components/NotificationTile'
 import { OthersLink } from './components/OthersLink'
 import { UserNameLink } from './components/UserNameLink'
-import { UserProfilePictureList } from './components/UserProfilePictureList'
 import { IconRepost } from './components/icons'
 import { entityToUserListEntity, USER_LENGTH_LIMIT } from './utils'
