+++ packages/web/src/components/user-profile-picture-list/UserProfilePictureList.module.css
@@ -62,4 +62,4 @@
   font-size: var(--harmony-font-s);
   font-weight: var(--harmony-font-bold);
   text-shadow: 0px 2px 4px rgba(0, 0, 0, 0.2);
-}
+}
