+++ packages/web/src/components/user-profile-picture-list/index.ts
@@ -1,0 +1,2 @@
+export { UserProfilePictureList } from './UserProfilePictureList'
+export type { UserProfileListProps } from './UserProfilePictureList'
